+++ packages/tdl/index.d.ts
@@ -82,6 +82,9 @@ export default Client
 export type LoginUser = {
   type: 'user',
   getPhoneNumber: (retry?: boolean) => Promise<string>,
+  getEmailAddress: () => Promise<string>,
+  getEmailCode: () => Promise<string>,
+  confirmOnAnotherDevice: (link: string) => void,
   getAuthCode: (retry?: boolean) => Promise<string>,
   getPassword: (passwordHint: string, retry?: boolean) => Promise<string>,
   getName: () => Promise<{ firstName: string, lastName?: string }>