+++ packages/tdl/src/prompt-stubs.js
@@ -4,6 +4,18 @@ export const getPhoneNumber = (): Promise<string> => {
   throw new Error('Default getPhoneNumber is not supported')
 }
 
+export const getEmailAddress = (): Promise<string> => {
+  throw new Error('Default getEmailAddress is not supported')
+}
+
+export const getEmailCode = (): Promise<string> => {
+  throw new Error('Default getEmailCode is not supported')
+}
+
+export const confirmOnAnotherDevice = (): void => {
+  throw new Error('Default confirmOnAnotherDevice is not supported')
+}
+
 export const getAuthCode = (): Promise<string> => {
   throw new Error('Default getAuthCode is not supported')
 }